requirements.txt
aniso8601==7.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
Flask==1.1.1
Flask-GraphQL==2.0.0
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
promise==2.2.1
python-dateutil==2.8.1
requests==2.22.0
Rx==1.6.1
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
yapf==0.28.0
